Borrow strength to improve prediction
This file contains predictions for the data of confirmed cases till 8/24/20.  This project is showing the approach when you have less data for training deep learning models. This approach shows that we can use the weights to gain more strength from the states that have some similarities may be the same neighborhood or rise and fall having the same pattern or any other similarities. Here NJ and NY are used to show how we can borrow some strength to boost the prediction.  NY's weight is used and applied to NJ's data, and then prediction and forecasting are made.